perf: run rebalancer simulations in parallel workers
