+++ typescript/rebalancer-sim/package.json
@@ -28,7 +28,7 @@
     "generate-scenarios": "tsx scripts/generate-scenarios.ts",
     "lint": "oxlint -c ../../oxlint.json",
     "format": "oxfmt --write ./src",
-    "test": "mocha --config .mocharc.json './test/**/*.test.ts' --exit",
+    "test": "mocha --config .mocharc.json './test/**/*.test.ts' --parallel --jobs 2 --exit",
     "test:ci": "mocha --config .mocharc.json './test/scenarios/**/*.test.ts' --exit"
   },
   "dependencies": {
